Refactor MAX_MOVE_X and MAX_MOVE_Y to use a single mutex
This commit refactors the code to use a single mutex for MAX_MOVE_X and MAX_MOVE_Y instead of two separate mutexes. This change has the potential to reduce the overhead of locking and unlocking mutexes and may improve performance. However, the performance gain from this may be negligible.